+++ index.js
@@ -3,8 +3,8 @@
 import React,{} from 'react';
 
 import {
-	PickerIOS,
 	Platform,
+	NativeModules
 } from 'react-native';
 
 import WheelCurvedPicker from './WheelCurvedPicker'
@@ -19,6 +19,14 @@ import DatePicker from './src/DatePicker';
 
 import DateRangePicker from './src/DateRangePicker';
 
+let PickerIOS = null;
+if(Platform.OS === 'ios') {
+	if(NativeModules.RNWPicker) {
+		PickerIOS = require('./src/PickerIOS').default;
+	} else {
+		PickerIOS = require('react-native').PickerIOS;
+	}
+}
 
 export {
 	PickerHeader,